Adjust logo to match grey background in jekyll-theme-hacker
I adjusted the background of the DisInfoSec logo to match the
background color of the web page, which uses jekyll-theme-hacker.
I further cropped the image so it would line up visually with the
page's text and ran the result through optipng.

I chose to adjust the background color rather than make it
transparent so that the result will still be visible on a white
background, e.g. at
https://github.com/Capricornucopia/DisInfoSec/blob/master/index.md